requirements.txt
alembic==1.5.2
Babel==2.9.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.10.1
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.5
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.22
urllib3==1.26.2
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
